RDKBWIFI-24: fix compilation issue (#124)
Reason for change: builds are failing
Test Procedure: XB devices should compile
Risks: Low
Priority:P0

+++ source/webconfig/wifi_encoder.c
@@ -457,9 +457,8 @@ webconfig_error_t encode_vap_common_object(const wifi_vap_info_t *vap_info,
     cJSON_AddBoolToObject(vap_object, "MboEnabled", vap_info->u.bss_info.mbo_enabled);
 
     memset(extra_vendor_ies_hex_str, 0, sizeof(extra_vendor_ies_hex_str));
-    for (int i = 0; i < vap_info->u.bss_info.vendor_elements_len; i++) {
-        sprintf(extra_vendor_ies_hex_str + (i * 2), "%02x",
-            vap_info->u.bss_info.vendor_elements[i]);
+    for (unsigned int i = 0; i < vap_info->u.bss_info.vendor_elements_len; i++) {
+        sprintf(extra_vendor_ies_hex_str + (i * 2), "%02x", (unsigned int) vap_info->u.bss_info.vendor_elements[i]);
     }
     cJSON_AddStringToObject(vap_object, "ExtraVendorIEs", extra_vendor_ies_hex_str);
